# backend/docs/swagger.yaml
basePath: /
definitions:
  admin.rehashRequest:
    properties:
      new_key:
        maxLength: 200
        type: string
      old_key:
        maxLength: 200
        type: string
    required:
    - new_key
    - old_key
    type: object
...
  model.Response:
    properties:
      code:
        description: 五位业务码，"00000" 表示成功
        type: string
      data: {}
      error:
        description: 业务错误文案（成功时为空串）
        type: string
    type: object
...
info:
  contact: {}
  description: 电子书阅读器后端 API（账号/认证/评论/操作日志）。业务响应用统一信封 {code,error,data}，HTTP 恒为 200；需认证的接口用
    Bearer access token。
  title: ebook-server API
  version: 1.0.0
paths:
  /admin/api/comments:
    get:
      parameters:
      - description: 内容关键字（模糊匹配）
        in: query
        name: keyword
        type: string
      - description: 书名过滤
        in: query
        name: book_name
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 后台评论列表
      tags:
      - 管理后台
  /admin/api/comments/{id}:
    delete:
      parameters:
      - description: 评论 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 后台删除评论（治理）
      tags:
      - 管理后台
  /admin/api/comments/rehash:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新旧聚合键
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.rehashRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 后台全局改评论聚合键（跨用户）
      tags:
      - 管理后台
...
  /admin/api/logs:
    get:
      parameters:
      - description: HTTP 方法（如 GET/POST）
        in: query
        name: method
        type: string
      - description: 路径片段（模糊匹配）
        in: query
        name: path
        type: string
      - description: 按用户 UID 过滤
        in: query
        name: user_id
        type: integer
      - description: 业务码（如 A0303）
        in: query
        name: error_code
        type: string
      - description: 只看失败请求：1 或 true
        in: query
        name: failed
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 后台操作日志列表
      tags:
      - 管理后台
...
  /admin/api/users:
    get:
      parameters:
      - description: 关键字（模糊匹配邮箱/用户名/昵称，纯数字兼命中 UID）
        in: query
        name: keyword
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 后台用户列表
      tags:
      - 管理后台
  /admin/api/users/{uid}:
    get:
      parameters:
      - description: 用户 UID
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 后台用户详情
      tags:
      - 管理后台
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"